Refactored pages dev and pages build
